+++ src/lib.rs
@@ -3,6 +3,7 @@
 mod canvas;
 mod dimension;
 mod error;
+mod frame;
 mod handle;
 mod pixels;
 mod texture;
@@ -10,6 +11,9 @@ mod texture;
 pub use canvas::*;
 pub use dimension::*;
 pub use error::*;
+pub use frame::*;
 pub use handle::*;
 pub use pixels::*;
 pub use texture::*;
+
+pub use wgpu;