+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Any notable changes to G2d will be listed in this file.
 ## [unreleased]
 ### Added
 - `WindowHandle` (A G2d `Handle` implementation for drawing to a window).
+- `Frame` API (for drawing to a window's surface).
+- Re-export `wgpu`
 
 ### Fixed